Databases: Support postgres schema in production environment.
- Unlike dev environment, schema will not be created if it does not
exist and instead will throw an error. The dev should also make sure
that they have sufficient usage and create priviliges on the schema.
- The env variables in .env file will not be picked up by the migration
script. Please export them explicitly.
- The migration script inside migration/sql was copied to the postgres
directory to accomodate the postgres schema. That migration will now
not run for both mssql and postgres.
- You might notice that we set the schema in the dataSource from the
env variable and we still modify the SQL strings in migrations file to
include schema name. This is because queryRunner when running
raw sql commands runs the commands directly without modification.
But we still need to set the schema because TypeORM does a select
operation on `_jackson_migrations` table and the schema name is
needed in the dataSource for that.