feat: Add deleteTracesByProduct method to AdminController (#2617)
* feat: Add deleteTracesByProduct method to AdminController

This commit adds a new method, deleteTracesByProduct, to the AdminController class in the npm/src/controller/admin.ts file. This method allows for the deletion of traces by product. It uses the getTracesByProduct method to retrieve traces in batches of 50 and deletes them using the tracerStore. The method is also implemented in the pages/api/v1/sso-traces/product.ts file as a DELETE handler.

* chore: Add DELETE handler for product API endpoint in product.ts file

* chore: Add DELETE handler for product API endpoint in product.ts file

* feat: Add countByProduct method to SSOTracer

This commit adds a new method, countByProduct, to the SSOTracer class in the npm/src/sso-tracer/index.ts file. This method allows for counting the number of traces by product. It uses the tracerStore's getCount method to retrieve the count based on the product name. The method is also implemented in the pages/api/v1/sso-traces/product/count.ts file as a GET handler.